+++ src/main/org/apache/groovy/internal/vmplugin/vm8/Java8.java
@@ -16,7 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.codehaus.groovy.vmplugin.vm8;
+package org.apache.groovy.internal.vmplugin.vm8;
 
 import org.codehaus.groovy.ast.AnnotationNode;
 import org.codehaus.groovy.vmplugin.v7.Java7;
